Refactor C++ code into modular files and switch to setuptools build system
Split the monolithic _pyensmallen.cpp file into modular components:
- utils.hpp: Common utilities and wrapper classes
- newton_type.hpp: L-BFGS implementation
- constrained.hpp: Frank-Wolfe algorithm for constrained optimization
- first_order.hpp: Adam and its variants

Also updated the build system to use setuptools instead of meson for better compatibility.
