+++ .github/workflows/build.yml
@@ -88,7 +88,7 @@ jobs:
       - name: Install Python dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install setuptools build meson meson-python pybind11
+          pip install setuptools build wheel pybind11
 
       - name: Build sdist
         run: python -m build --sdist --outdir wheelhouse/
