+++ meson.build
@@ -10,7 +10,7 @@ armadillo_dep = dependency('Armadillo')
 
 # Simplify to bare minimum
 py.extension_module('_pyensmallen',
-    'pyensmallen/_pyensmallen.cpp',
+    'pyensmallen/module.cpp',
     dependencies: [pybind11_dep, armadillo_dep, ensmallen_dep],
     install: true,
     install_dir: py.get_install_dir() / 'pyensmallen'
@@ -18,7 +18,14 @@ py.extension_module('_pyensmallen',
 
 # Install Python sources separately
 py.install_sources(
-    ['pyensmallen/__init__.py', 'pyensmallen/losses.py'],
+    [
+        'pyensmallen/__init__.py',
+        'pyensmallen/losses.py',
+        'pyensmallen/utils.hpp',
+        'pyensmallen/first_order.hpp',
+        'pyensmallen/newton_type.hpp',
+        'pyensmallen/constrained.hpp'
+    ],
     pure: false,
     subdir: 'pyensmallen'
 )
